app/requirements.txt
psycopg2-binary==2.8.6
Pillow==8.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
django-filter==2.4.0
django-ckeditor==6.1.0
drf-yasg==1.20.0
autocrop==1.1.0